# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hydraulic Monitoring System API",
        "title": "TP-API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/boxes/{id}/records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boxes"
                ],
                "summary": "List sensor records for a box",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Box ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Min timestamp (seconds)",
                        "name": "time_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max timestamp (seconds)",
                        "name": "time_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaginatedResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boxes"
                ],
                "summary": "Add a sensor record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Box ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Record data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Record"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/boxes/{id}/reports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boxes"
                ],
                "summary": "Generate daily report for a box",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Box ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Min timestamp (seconds)",
                        "name": "time_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max timestamp (seconds)",
                        "name": "time_max",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DailyReport"
                            }
                        }
                    }
                }
            }
        },
        "/data/box/{box_id}/count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data"
                ],
                "summary": "Count sensor records for a box",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Box ID",
                        "name": "box_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Min timestamp (seconds)",
                        "name": "time_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max timestamp (seconds)",
                        "name": "time_max",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/groups/{id}/records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List sensor records for all boxes in a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Min timestamp (seconds)",
                        "name": "time_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max timestamp (seconds)",
                        "name": "time_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaginatedResponse"
                        }
                    }
                }
            }
        },
        "/groups/{id}/records/latest": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "List sensor records latest for all boxes in a group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaginatedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DailyReport": {
            "type": "object",
            "properties": {
                "avg": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "max": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "min": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                }
            }
        },
...
        "domain.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/domain.PaginationMeta"
                }
            }
        },
        "domain.PaginationMeta": {
            "type": "object",
            "properties": {
                "filter": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Record": {
            "type": "object",
            "additionalProperties": true
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for JWT authentication (format: Bearer \u003ctoken\u003e)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}